Use dataframe index as the x axis in charts.
In the process, I also unified DEFAULT_COMPONENTS and
REQUIRED_COMPONENTS in charts. This simplifies things, even though the
end result is slighly less customizeable charts. For instance, now every line
chart will have the x_axis component set, and there's no way to unset
that.

For now, this should be more than good. But we need to think about how
much customizability we want for the future.