Replace requirements.txt with pipenv in Makefile.
Originally there was install_requirements.in and requirements.in which
then used pip-compile to generate install_requirements.txt and
requirements.txt

That was all deleted in
https://github.com/streamlit/streamlit/commit/b52bc73e046a8dd35856a096ee438cb30e79fad6#diff-0ef0eb202fb86cb8145a10b267cba8c9

Nothing was really pinned in the various *.in files, they're only pinned
in the *.txt files.  pipenv does the same thing but with less files.