Add user-tests directory and setup some basic scaffolding for non unit tests.
I know there's an examples and admin already.  This uses the pytest
framework which isn't being heavily used in this example but will be
important later on for the ability to easily reuse mocks, etc.

While this test is trivial, it isn't a unit test.  It actually goes and
uses the internet to check for an ip address which is an integration
test.  We will eventually test more user specific things.