Check version on startup (#106)
Occasionally check the installed streamlit version against the latest in PyPI, and show a notice if there's a new version available.

The check runs randomly with a 5% chance, so that we're not spamming the user frequently.

I used `requests` (instead of urllib) for this, which entailed adding it as an explicit dependency in our Pipfile. All the other places in streamlit that used urllib for fetching over HTTP have been replaced with requests. (And there are some additional tests that didn't previously exist.)

Fixes #65