Handle lack of trailing slash (#528)
* Make the trailing slash optional

Add a kw-arg to make_url_path_regex that controls whether an optional
trailing slash is included at the end of the path pattern. This will be
used to generate a path without the trailing slash for redirects.

* Create a handler that adds a slash

Create a simple request handler that simply uses the
tornado.web.addslash decorator to add a slash to the end of the request
path for GET requests.

* Redirect to base path with slash

If a request comes in for the app on the base path without a trailing
slash, respond with a redirect to the path with the trailing slash.
Otherwise, users just receive a 404 Not Found.

* Add test case for trailing slash

Add an extra test case ensuring that trailing slashes are removed.

* Gather keyword args

Python 2.7 doesn’t support having keyword args after an args collector.
Replacing trailing_slash with gathering all keyword args, and then
extract it from there.