Workaround for deep import reloads not working (#537)
**Issue:** https://github.com/streamlit/streamlit/issues/366

Fixes #366, #358

**Description:** 
This is a sketch for now, we may want to look at a more sophisticated fix. Although it does fix the linked issue, it does so in a somewhat scorched-earth manner.

Essentially what is happening is that if you have a module which is not directly imported by the base module (e.g.` app.py` imports `a.py`, `a.py` imports `b.py`), then changes to that module will not be reloaded when a file change is detected.

As far as I can understand, this is what is happening:
- User changes `b.py`
- Reload logic:
  - Deletes `b` from `sys.modules`
  - Calls eval on the code from `app.py`
  - The code in `app.py` imports `a`, which is already present in `sys.modules`, and so we are done
  - The result is that we keep using the old copy of `b` in memory until `a` is reloaded

The ideal solution would be to unload both the modified module, as well as any modules that are importing it, but determining this tree of all possible paths for importing a given module may not be worth the effort.

Instead, what this PR does is simply unloads all watched modules. So in the scenario above, `a` gets unloaded as well when `b` changes, and so at the eval point both modules get reloaded, therefore reflecting the changes to `b` in the running app.

---

**Contribution License Agreement**

By submiting this pull request you agree that all contributions to this project are made under the Apache 2.0 license.