Fix the "keyPrefix" option in static report sharing (#724)
This fixes an issue where using a `keyPrefix` with a `/` in it would result in the frontend failing to download a statically shared report.

The issue was pretty simple - we were just doing some naive string parsing in `ConnectionManager` and `StaticConnection`. This fix implements the correct parsing in a new function called `getBucketAndResourceRoot`; both `ConnectionManager` and `StaticConnection` use this new function instead of implementing the (broken) logic themselves.

UNRELATED to this PR are a handful of `pylint` fixes; CircleCI was failing my build because of some formatting in a few python files. They had to come along for the ride to make Circle happy :/

Fixes #710