Pass --skip-lock to pipenv under pipenv-install Makefile target.
By not passing this, we were violating the "doesn't update the
Pipfile.lock." comment on this Makefile target. I don't see where the
lockfile was later used, and so was most likely providing little
benefit.

Finally, invocation before this change took ~1m13s on my machine when
Pipfile.lock wasn't present, and ~33s when it was. With this change,
it's reliably ~17s.

The big question is if the lockfile was providing any value that I
missed picking up, so that's the major review question.

CC @tconkling