Version 0.61.0 (#1512)
* Enable WebSocket compression (#1506)

Fixes #463

From a lines-of-code perspective, this is a trivial change; Tornado supports WebSocket compression out of the box, and this PR just flips the switch to enable it.

(This obviously adds a bit of memory and CPU overhead to each WebSocket connection.)

* Fix DeckGL View State (#1481)

* Merge server and client view state changes

* Version 0.61.0

* Update terminal welcome + telemetry messages (#1510)

* Up resource class used for Python 3.8 tests in CircleCI (#1520)

* Up resource class used for Python 3.8 tests in CircleCI

* Ooops resource classs change should have be to the "cypress" flow

* Update notices.

* Fix mypy in CircleCI (#1519)

Our `mypy` task has been failing in CircleCI.

The issue is that `pipenv` is failing to deterministically resolve dependency versions, and is frequently resolving our `tornado=>=5.0, <6.0` dependency to Tornado 6 (because of other transitive dependencies on Tornado).

The solution is to use the `--sequential` flag in our `pipenv install` build step. From the pipenv docs:

> Installation is intended to be as deterministic as possible — use the --sequential flag to increase this, if experiencing issues.

So: pipenv installation is intended to be as deterministic as possible, but it's also possible to "increase its determinism" with a flag! (I think this is...madness.)

This PR *also* cleans up and documents a few things that I found confusing while tracking this down (with @jrhone's help!):

- The `mypy` cache is now stored in the project directory (and gitignored), rather than in `/tmp`
- We expire our CircleCI virtualenv cache key every day; I documented why we do this.
- Added docs to the "Create virtualenv" CircleCI build step.
- Removed the protobuf-caching step from the CircleCI jobs, because it's not actually doing anything.
- Updated the mypy ini to target Python 3.8, and fixed a new mypy warning.

* Make credentials prompt align to 80-char max, and do some tiny cleanups.

* Lint

* Changelog

* Remove "mapbox token" breaking change notice from changelog, since I dropped that PR for this launch

Co-authored-by: Tim Conkling <tconkling@gmail.com>
Co-authored-by: Jonathan Rhone <rhone.j@gmail.com>
Co-authored-by: karrie <karrie@streamlit.io>