Require tornado >= 6, instead of 5
Because we're now typechecking our tornado usages with mypy, we're
*not* actually compatible with tornado 5 (at least as far as
typechecking is confirmed). That is, if you end up with tornado 5
in your venv, `make mypy` will (rightly) have a bunch of failures.

Since tornado 6 changes APIs, and since we're not actually testing
against both tornado 5 and 6 compatibility, we shouldn't allow for
tornado < 6. (Or: if we *do* want to support tornado 5, we need to
actually test against it.)