Upgrade react-scripts and TypeScript (#1677)
- react-scripts: `3.0.1` -> `3.4.1`
- TypeScript: `3.6.3` -> `3.8.3`

The eslint version we require is part of react-scripts, so we don't need an exact dependency on it. (Removing the explicit eslint dependency makes for easier upgrading of react-scripts.)

Most of the changes in this PR are to fix eslint errors and warnings that this new version of eslint catches.