Components: allow fullscreen content (#1713)
Adds "fullscreen" to our `iframe allow` feature policy. This applies to `components.iframe`, `components.html`, and `components.declare_component` components. 

With this attribute, components can use the `Element.requestFullscreen()` web API.

(We omitted this simply because we hadn't made a call on it. The team discussed it today, and we're all cool with allowing it.)

Fixes #1710