Update DateInput to handle dates below 100
The Date constructor makes assumptions based on the input string that are
inconsistent. When the year is set to below 100, it assumes the year is a
suffix of 19 or 20. (e.g. '0099' returns 1999 year and '0001' returns 2001
year).

This change avoids it by using moment to parse a specific syntax (YYYY/MM/DD)