WidgetStateManager.getOrCreateWidgetStateProto -> createWidgetStateProto
This fixes a bug in WidgetStateManager - previously, changing the type of a WidgetState's oneOf value would *not* remove its previous value. In the case of `bytesValue`, its `value` string was also not properly updated.

(It's not clear to me whether this is a bug in protobufjs - it seems like a setter function should be responsible for updating things appropriately? But no such setter function is called in these circumstances - it's just raw data being set.)