Components: support for bytes (#1981)
Adds support for sending bytes to, and receiving bytes from, components

**Python -> Frontend**:
- In Python, Component functions can now take `bytes` and `bytearray` values as arguments.
- These are delivered to the frontend in the `special_args` list inside `ComponentInstance.proto`. (DataFrames are also sent inside this list.)
- The frontend delivers these bytes args to the iframe as `UInt8Array` values (which is their "native" datatype within protobuf-js).

**Frontend -> Python**:
- The `Streamlit.setComponentValue` function (implemented in our component-lib npm package) can now take `ArrayBuffer` values, and any TypedArray value (`UInt8Array`, `UInt16Array`, etc etc). These are all converted to `UInt8Array` and shipped over to Python, where they're deserialized as `bytes`.

There's also a fix to `WidgetStateManager` in here: `WidgetStateManager.getOrCreateWidgetStateProto` is now just `createWidgetStateProto`, and it always overwrites any existing value. Previously, changing the type of a WidgetState's oneOf value would *not* remove its previous value. In the case of `bytesValue`, its `value` string was also not properly updated.

(It's not clear to me whether this is a bug in protobufjs - it seems like a setter function should be responsible for updating things appropriately? But no such setter function is called in these circumstances - it's just raw data being set.)