WidgetStateManager: don't return `Long` for int values (#2104)
Our "intValue" and "intArrayValue" widget protobuf values represent values with sint64, because sint32 is too small to represent the full range of JavaScript int values. 

Protobufjs uses `number | Long` to represent sint64. However, we're never putting Longs *into* int and intArrays - because none of our widgets use Longs - so we'll never get a Long back out.

This changes the return value of `getIntValue` and `getIntArrayValue` to `number`, and asserts that a `Long` is never returned.

There's also a couple simple tests that verify we can pass MIN_SAFE_INT and MAX_SAFE_INT to these functions.