Revert BokehChart to use class-based components
This should fix the Bokeh duplication issue, which was introduced by the migration from class-based React component to function component.

Hypothesis is that useEffect() is run too often (specifically, it's also run on componentWillUnmount). This exposes a race condition: BokehEmbed.embed_item, which is actually a Promise, may run multiple times, causing two Bokeh charts to appear.

Sadly, React's useEffect() hook does not support async/await, so it's not easy to coordinate between two runs of updateChart.

It might be possible to use Hooks with a proper choice of optimizations (aka the second parameter to useEffect, the array of variables to skip), but the safe path is to revert for now.