Add draft of documentation for Theming
This is a rough draft as it's likely that enough things will change as a
result of the Product Launch Review that some sections will have to be
rewritten.

Reworks in response to feedback probably won't be major changes, however, so
it'd still be useful to get some eyes on this now.