Stop warn logs from being spammed every time a checkbox is rendered (#2938)
This slipped by when the design pass PR was merged as I guess I never
had a browser console open at the right time :cry:

I happened to notice a ton of warn logs being spammed to the browser
console whenever a checkbox was rendered. Unfortunately, this seems to
be due to the incompatibility of the short/long-hand versions of these
names and the fact that baseweb uses the long versions for some
components, so we don't have much of a choice other than to be very,
very verbose as well.