Downgrade decamelize to version 4.0.0
Apparently version 5.0.0 breaks in Safari (and the author has no plans of
fixing this as it breaks other features), but the humanize-string
package that relies on it only needs a version more recent than 2.0.0,
so we can downgrade this here.

Ideally, we may want to just get rid of the need for humanize-string
entirely in the near-ish future since its usage is pretty minimal.