Force file_watcher_type to "poll" for secrets.toml
Streamlit Sharing makes an app's 'secrets.toml' file available from a special
file system that is incompatible with watchdog. So by default, in Sharing,
changes to secrets.toml won't be detected at runtime.

With this change, we force the use of PollingFileWatcher (for secrets.toml only)
instead of using whatever file_watcher method is set in config.