Only use localStorage when it is available
We ran into an issue with theming on Streamlit Sharing on Chrome
Incognito mode because streamlit apps are run in an iframe so are
considered third party code.

Chrome Incognito disallows third party cookies/data by default, which
was causing the app to fail to load in this case. A (temporary) solution
is to simply avoid trying using `localStorage` when it's unavailable to
us.