Only use localStorage when it is available (#2974)
* Only use localStorage when it is available

We ran into an issue with theming on Streamlit Sharing on Chrome
Incognito mode because streamlit apps are run in an iframe so are
considered third party code.

Chrome Incognito disallows third party cookies/data by default, which
was causing the app to fail to load in this case. A (temporary) solution
is to simply avoid trying using `localStorage` when it's unavailable to
us.

* Remove leftover .only

* Grab localStorage in try block too

* Include pathname in key for cached theme

This is necessary as a browser's localStorage is scoped to a domain,
so all apps on Streamlit sharing will have the same key for their cached
theme (so will overwrite each other) if we don't do this.