Allow users to click and drag in the color picker
It turns out that doing this was actually quite easy as we only need to
use `react-color`s `onChange` instead of `onChangeComplete`. Note that the
corresponding event fires a lot when clicking and dragging, but the
`onChange` handler passed to the ColorPicker component is only called
when the picker popover is closed.