Stop warn logs from being spammed in dev mode due to override names (#2994)
I noticed a bunch of warn logs in my browser console every time a tooltip is
rendered, and it turns out the reason for this was essentially the same
reason from when this appeared last / was solved by #2938.

This doesn't affect non-dev setups, so I'm not sure how big of a deal this
actually is and whether we should make it a convention to always use the
long names in BaseWeb overrides.