Add snippets about CSS vars and minimum streamlit-component-lib versions (#3008)
The referenced GitHub issue points out that we don't have any examples
for how to actually use the CSS vars that we provide to custom
components. We actually have a snippet about this in our theming launch
blog post, so this commit just adds that bit here verbatim.

Additionally, this adds a note about the minimum supported
streamlit-component-lib version for theming compatibility. This feels a
bit weird to have to do since our docs are versioned, but I think it's
needed in this case since the streamlit-component-lib package version is
decoupled from the PyPI package version that the docs are tied to.

Fixes #2996