Downgrade st.warning in local_sources_watcher to a log
I ran into some trouble trying to call st.warning outside of code run by
the ScriptRunner, which made me remember that some code changed in a
recent PR (#3001) did just that.

I double-checked that this code indeed doesn't work as calls that
end up sending Delta protos rely on a ReportContext, which isn't
available in the LocalSourcesWatcher.

Since we can't present the warning in Streamlit itself (which would be
ideal), I just downgraded this to be a warning log for now.