Upgrade bokehjs to 2.2.2, and run `npx browserslist@latest --update-db` (#3090)
- bokehjs: 2.0.0 -> 2.2.2
- ran `npx browserslist@latest --update-db`

This is not a solution for https://github.com/streamlit/streamlit/issues/2156. The latest bokehjs is 2.3.1, but using bokehjs 2.3.0 or 2.3.1 gives a runtime error when the chart is rendered: 

"ReferenceError: Cannot access ‘ColumnarDataSource’ before initialization".

It looks like this is a cyclic dependency issue, and that it's been addressed by bokehjs here: https://github.com/bokeh/bokeh/pull/11140. Presumably, when bokehjs 2.4 releases, we'll be able to it.

(More context in the bokehjs forums, here: https://discourse.bokeh.org/t/problem-importing-bokeh-bokehjs-module-in-javascript-app/7510/12)