Fix a (dev-mode-only) theming bug with cached themes
There's a cached theme bug that should only be possible to run into in dev
mode that frustrated me enough that I still want to fix it. It's also
probably possible to run into in between the react app loading and the script
run starting, but in practice the delay between that happening is short
enough to make that really hard.

The issue is that if you're connected to the react dev server, but the
streamlit server isn't running, a cached custom theme may be loaded as
the active theme of the app, but it won't appear as an available theme
in the selector menu.

This is just because the default value for that piece of react state
doesn't consider that there may be a custom theme option, so we just
need to add it on the ThemedApp component's first render.