Merge feature/theming-fast-follow into develop (#3164)
* Add support for base and partial themes (#3049)

We want users to be able to specify a partial theme that inherits from the
preset light/dark themes.

To do this, this PR

* Adds a new "base" theme config option that can be either "light" or "dark".
  The option defaults to "light" if unspecified.
* Removes the validation requiring that a theme be fully specified. Now, a
  theme is sent to the client if at least one theme option is set.
* On the client, uses the given "base" theme config option.

* Don't accidentally mutate the light/dark theme when creating a new one (#3078)

It turns out that we were accidentally mutating the "starterTheme"'s
`emotion.inSidebar` field since the spread operator only does a shallow
copy, and `lodash/merge` mutates the target object. This resulted in a
spacing bug where, when a custom theme is set, the base theme for it
(which depends on the background color of the custom theme) has weird
spacing since it has `emotion.inSidebar` set to true for it, even in the
main content area.

* Drop size of cached themes and recompute full themes from them (#3091)

* Reorder tests and improve test descriptions in theme/utils.test.ts

* Drop size of cached themes and recompute full themes from them

This fixes two issues that currently exist with cached themes that don't
currently manifest, but would eventually be problematic if left unfixed.
  1. We cache the full theme, which is huge (order of tens of kilobytes)
     compared to the minimal representation of the actually-configurable
     parts of a theme. This change only stores the configurable pieces
     and recomputes the rest, which shrinks the size of what we're
     storing to a few hundred bytes.
  2. Newly added derived theme colors may not exist in a cached theme,
     which causes weird visual bugs until the theme is changed and
     derived colors are recomputed. This is no longer an issue since
     we're storing a minimal representation of a cached theme and
     recomputing the rest.

* Give people (mostly) minimal theme toml definitions (#3103)

Now that we can define themes by specifying a base theme (or having the
base theme default to "light") and selectively overwriting options,
there are often multiple ways to define the same theme. This change
improves the "Copy theme to clipboard" button to give the user a smaller
toml snippet defining their theme when possible.

The toml snippets now:
1. Don't specify a font unless it's changed from "sans serif"
2. Set a "light" or "dark" base (depending on what the currently defined
   custom theme is closer to) and partially define the changed
   properties when doing so produces a minimal config.
3. May not set a base at all if enough colors are changed so that doing
   so is not needed.

The reason we give people "mostly" minimal theme toml definitions is
that we include `base="light"` even when it's not strictly required to
be more explicit.

* Fix a (dev-mode-only) theming bug with cached themes (#3109)

There's a cached theme bug that should only be possible to run into in dev
mode that frustrated me enough that I still want to fix it. It's also
probably possible to run into in between the react app loading and the script
run starting, but in practice the delay between that happening is short
enough to make that really hard.

The issue is that if you're connected to the react dev server, but the
streamlit server isn't running, a cached custom theme may be loaded as
the active theme of the app, but it won't appear as an available theme
in the selector menu.

This is just because the default value for that piece of react state
doesn't consider that there may be a custom theme option, so we just
need to add it on the ThemedApp component's first render.

* A bit of theme caching and theme toml generation polish (#3107)

* Make minimal theme toml generator case insensitive

Also changed colors.white to be lowercase to match the case of everything else.

* Also remove cached themes with key "stActiveTheme"

Cached themes in pre-release versions of theming had this key, which I forgot
about while writing the original PR adding cached theme cleanup.

* Tweak theme colors so slider troughs always have enough contrast. (#3118)

* Tweak theme colors so sliders troughs always have enough contrast.

This also changes how st.number_input looks. Now the +/- buttons have
the same color as the text-based input area.

* Remove background from pre/code inside alert elements

* Fix shadow color

* Add border around code in markdown inside alertbox

* Update non-cypress snapshots

* Take snapshots of focused checkbox and radio widgets

* Add new and update existing snapshots

Co-authored-by: Vincent Donato <vincent@streamlit.io>

* Revise wording for base description and warning messages (#3171)

I didn't put much thought into the copy that was originally here. The
new versions are revised after some discussions with product on what the
description / warning messages should be.

Co-authored-by: Thiago Teixeira <thiago@streamlit.io>