Temporarily disable extremely flaky tests
I added some snapshot tests for things that previously didn't have any
test coverage in the theming fast follow feature branch. These tests
seemed okay in the branch but suddenly turned flaky to the point of
being unusable once it got merged into develop, so this commit disables
them for now to un-break the build until the tests can be de-flaked and
revived.