Revive the docker image used to take cypress snapshots
This should make it possible to generate and update snapshots without relying
on CI in some way (either by deleting snapshots to have CI regenerate them or
having CI generate diffs to be processed by the cropping script)

After this PR goes in, I'll spend some time reworking our wiki pages to include
instructions on how to use this image as well as an overview of all of the
available options for updating snapshots.

Unfortunately, this isn't a perfect cure for our snapshot woes for a few
reasons.

* it takes *forever* to build image, mostly due to the length of time it
  takes to install Python dependencies. As far as I can tell, this is
  naturally quite slow, and there's not much we can do about it.
* for whatever reason I'm having trouble getting component-template
  tests to work, but this is just a minor annoyance for now as those
  don't change much. This does mean that you can't run all e2e tests
  easily since they'll fail midway through due to these tests blowing
  up, but doing this takes so long that it seems unlikely that it'll be
  done often anyway.