Add support for toml syntax highlighting in code blocks (#3140)
We changed the behavior of auto-syntax highlighting in code blocks in #2814 to
*not* attempt to highlight code as if it were python if the language
specified isn't a supported one, which had the side effect of removing
syntax highlighting from toml code blocks (toml was being highlighted as if it
were python). This change adds toml highlighting back using prismjs' native
support.

We also re-order and re-group the imports in the file.