Implement _get_session_state
_get_session_state fetches the SessionState object for the session it is called
from.

We're once again mostly able to copy the implementation in the prototype for
this PR. The only notable change was to account for the fact that the
st.session_state API no longer allows users to initialize default values
for SessionState objects on creation (more precisely, users can no
longer directly call the SessionState constructor), so we can create a
SessionState object when we initialize a ReportSession.

We also had to

* Create an __init__.py file in the new state directory. We didn't
  notice that this hadn't been done before because mypy didn't start
  complaining until we started importing streamlit.state.session_state
  in other files.
* Move some test util code out of server_test.py and into
  server_test_case.py.