Expose st.session_state to users and add e2e tests
As with the past couple PRs, we were able to essentially just use the
prototype implementation here and improve its test coverage.

The only notable change is that we defined the LazySessionState class in
state/session_state.py so that the only thing done in
lib/streamlit/__init__.py is importing LazySessionState and creating an
instance of the class to be exposed to the user.