Move WidgetStateManager from ScriptRunner -> ReportSession
In the pre-session-state world, we were able to simply create a new
WidgetStateManager for each script run since widget state was only
stored on the client. In this world, WidgetStateManager could have
probably been replaced by some stateless helper functions as a new one
would be created for each script run.

Now that we're unifying widget state and session state, we need to
use WidgetStateManager to actually preserve state, so it can no longer
live in the very transient ScriptRunner class. Naturally, we move it to
the ReportSession class since we need it to interact with SessionState.