st.form: clear_on_submit (#3287)
Implements `clear_on_submit`, a new optional parameter for st.form. When a form has this flag, it means that the frontend "resets" all the form's widgets when the form is submitted.

CustomComponents are _not_ currently affected by this flag (this is by design - properly handling this would require a CustomComponents API update, and all widget-like components in the wild would need updating. So we're punting on it for now).

- `FormsManager` is now folded into `WidgetStateManager`.
- When a form is submitted, `WidgetStateManager` emits a signal. All widgets in the form respond to this signal by resetting their values locally and committing those values back to the WidgetStateManager.
- This resetting only happens on the frontend - the server doesn't know that values have been reset (this is by design).
- `FormClearHelper` is a utility class that implements a `useEffect`-like pattern for subscribing to the `formCleared` events. (None of our widgets are currently implemented using React hooks, so this is a workaround for easily and correctly composing that behavior into each widget.)
- Each widget has a new test for `clearOnSubmit` behavior.
- `st.form_submit_button` now takes an optional "help" parameter.

(Contains changes from https://github.com/streamlit/streamlit/pull/3286, which should be reviewed first)