Remove the `which-browser` dependency
We're currently using this lib to check if the user is on a browser that doesn't support screencasting, based on https://caniuse.com/?search=getdisplaymedia. But:

- we've failed to update this whitelist (Safari now supports getDisplayMedia, for example - so we're warning Safari users against this feature unnecessarily)
- which-browser is no longer being updated, and the dev actively recommends against using it
- which-browser downloads its fingerprinting data from an unsecured S3 bucket (it allows writing and deleting w/o creds)
- latacora recommended we remove it

We are still testing for the _existence_ of the `getDisplayMedia` function, and warning in its absence. This isn't 100% (some versions of browsers have a `getDisplayMedia` implementation that always returns an error-ing promise). We also catch initialization errors and warn when they happen.

(This PR also has some minor cleanup of withScreencast, including logging of uncaught "stopRecording" errors)