Remove the `which-browser` dependency (#3323)
We're currently using this lib to check if the user is on a browser that doesn't support screencasting, based on https://caniuse.com/?search=getdisplaymedia. But:

- which-browser is no longer being updated, and the dev actively recommends against using it
- which-browser downloads its fingerprinting data from an unsecured S3 bucket (it allows writing and deleting w/o creds)
- latacora recommended we remove it

This PR removes `which-browser`, and updates the screencast compatibility check to rely solely on the presence of the feature. It adds an explicit check for webm support, which properly detects that Safari is unsupported.

(This PR also has some minor cleanup of withScreencast, including logging of various errors that make debugging easier)