Add all callback-related args to all widgets (#3375)
* Add all callback-related args to all widgets

Note that we don't actually do anything with these args aside from
verify that the widget is well-behaved and follows some rules we have
around callbacks. Actually hooking them in can happen once session and
widget state are more properly unified, and we might as well get this
routine work out of the way now.

I hesitantly but intentionally omitted a good amount of type info from
both the docstrings and function signatures. This feels a bit weird, but
I think it's necessary because

  * when sphinx/RTD renders documentation, it uses the function
    signature to do so, and unfortunately things look pretty awful for
    long function signatures with complex types.
  * existing documentation in the API Reference does not state that some
    args are optional in the type but does do so in the following
    explanatory text, and I didn't want to take the time to also rewrite
    parts of the docstrings for every widget.

We can probably look to resolve these issues later, but I think doing so
is outside of the scope of this work on session state.

* Change "color picker" -> color_picker in docstring

* Readd is_new_value session_state method

We readded the method under a slightly different name is_new_state_value.
This was removed in a PR probably because it was unused/seemed like dead
code, but we need it to enforce some session_state API rules.