Maintain state priority ordering when compacting
Previously, there was an issue with widget updates from the frontend
being lost sometimes, which was mitigated by compacting in multiple
places. One of the (seemingly redundant) compaction calls was removed,
which caused the bug to resurface. This is a proper fix for the bug,
which was the result of incorrectly preferring widget state over values
from session state set during callbacks.