Implement serde for file uploader widget
The file uploader is weird, because the values it communicates with the
frontend are not directly translatable into regular python values. They
must be used to look up the file information from the ReportContext.
This unfortunately results in deserialization requiring the widget id,
which no other widget requires, so they all need to take and ignore this
second parameter.

Several tests for file uploader worked by patching `register_widget`,
which now returns the deserialized values instead of leaving some work
to the widget function. These tests will need to be rewritten to work
for this, but I wanted to get the changes up for review first, since it
might be involved.